[h&s poc] refactoring with Storm decouple principle
