+++ docs/design/hub-and-spoke/v7/poc/wfm/src/main/java/org/openkilda/IFlowCrudCarrier.java
@@ -1,0 +1,11 @@
+package org.openkilda;
+
+import org.openkilda.model.FlCommand;
+
+public interface IFlowCrudCarrier {
+    void registerCallback(String key);
+    void cancelCallback(String key);
+    void installRule(String key, FlCommand command);
+    void checkRule(String key, FlCommand command);
+    void response(String key, String response);
+}
